autorizzazione e autenticazione implementati

+++ PasswordGen/Service/UtenteService/UtenteService.cs
@@ -10,8 +10,9 @@ namespace PasswordGen.Service.UtenteService
         {
             Db = db;
         }
-        public async Task<bool> NewUtente(string username, string password)
+        public async Task<bool> NewUtente(string? username, string? password)
         {
+            if (username is null || password is null) return false;
             if ((await Db.GetUtente(username, password)) is null)
                 if (Utente.Create(username, password) is Utente u)
                     return await Db.NewUtente(u);
@@ -19,20 +20,23 @@ namespace PasswordGen.Service.UtenteService
             return false;
 
         }
-        public async Task<bool> ChangeUtente(string username, string password, string? usernameNew, string? passwordNew)
+        public async Task<bool> ChangeUtente(string? username, string? password, string? usernameNew, string? passwordNew)
         {
+            if (username is null || password is null) return false;
             if ((await Db.GetUtente(username, password))?.ChangeCredenziali(usernameNew, passwordNew) is true)
                  return await Db.Save();
             return false;
 
         }
 
-        public async Task<bool> DeleteUtente(string username, string password)
+        public async Task<bool> DeleteUtente(string? username, string? password)
         {
+            if (username is null || password is null) return false;
             return (await Db.DeleteUtente(username, password));
         }
-        public async Task<Utente?> GetUtente(string username, string password)
+        public async Task<Utente?> GetUtente(string? username, string? password)
         {
+            if (username is null || password is null) return null;
             return (await Db.GetUtente(username, password));
         }
     }
